Move MoE routing-gradient tests into tests/test_tuner_trainer.py
Per review: match the test placement used by #1787 instead of a separate
test file.
