+++ tests/test_tuner_trainer.py
@@ -108,6 +108,87 @@ class TestTunerTrainer(unittest.TestCase):
         )
         self._assert_backward(model)
 
+    def test_granitemoe_backward(self):
+        from mlx_lm.models import granitemoe
+
+        model = granitemoe.Model(
+            granitemoe.ModelArgs(
+                model_type="granitemoe",
+                hidden_size=16,
+                num_hidden_layers=1,
+                intermediate_size=32,
+                num_attention_heads=4,
+                rms_norm_eps=1e-6,
+                vocab_size=32,
+                logits_scaling=1.0,
+                attention_multiplier=1.0,
+                embedding_multiplier=1.0,
+                residual_multiplier=1.0,
+                max_position_embeddings=128,
+                num_key_value_heads=2,
+                attention_bias=False,
+                rope_theta=10_000.0,
+                num_local_experts=4,
+                num_experts_per_tok=2,
+            )
+        )
+        self._assert_backward(model)
+
+    def test_granitemoehybrid_backward(self):
+        from mlx_lm.models import granitemoehybrid
+
+        model = granitemoehybrid.Model(
+            granitemoehybrid.ModelArgs(
+                model_type="granitemoehybrid",
+                vocab_size=32,
+                hidden_size=16,
+                intermediate_size=32,
+                num_hidden_layers=1,
+                max_position_embeddings=128,
+                num_attention_heads=4,
+                num_key_value_heads=2,
+                attention_bias=False,
+                embedding_multiplier=1.0,
+                attention_multiplier=1.0,
+                logits_scaling=1.0,
+                residual_multiplier=1.0,
+                layer_types=["attention"],
+                rms_norm_eps=1e-6,
+                rope_theta=10_000.0,
+                num_local_experts=4,
+                num_experts_per_tok=2,
+                shared_intermediate_size=32,
+            )
+        )
+        self._assert_backward(model)
+
+    def test_lfm2_moe_backward(self):
+        from mlx_lm.models import lfm2_moe
+
+        model = lfm2_moe.Model(
+            lfm2_moe.ModelArgs(
+                model_type="lfm2_moe",
+                vocab_size=32,
+                hidden_size=16,
+                intermediate_size=32,
+                moe_intermediate_size=16,
+                num_hidden_layers=2,
+                num_experts=4,
+                num_experts_per_tok=2,
+                norm_topk_prob=True,
+                num_attention_heads=4,
+                num_key_value_heads=2,
+                max_position_embeddings=128,
+                use_expert_bias=False,
+                num_dense_layers=1,
+                norm_eps=1e-6,
+                conv_bias=False,
+                conv_L_cache=3,
+                layer_types=["conv", "full_attention"],
+            )
+        )
+        self._assert_backward(model)
+
 
 if __name__ == "__main__":
     unittest.main()
